make more modules optional and update ci...
