+++ jumpscale/clients/docker/__init__.py
@@ -1,3 +1,6 @@
+__optional__ = True
+
+
 def export_module_as():
 
     from jumpscale.core.base import StoredFactory
