+++ jumpscale/data/fake/__init__.py
@@ -1,3 +1,6 @@
+__optional__ = True
+
+
 def export_module_as():
     from faker import Faker
     import sys
