+++ pyproject.toml
@@ -8,12 +8,15 @@ license = "MIT"
 
 [tool.poetry.dependencies]
 python = "^3.6"
-pdoc3 = "^0.6.3"
+importlib-metadata = "^3.4.0"
 pytoml = "^0.1.21"
 secretconf = "^0.1.2"
 better-exceptions = "^0.2.2"
+pycparser = "^2.20"
+zipp = "^3.4.0"
 pynacl = "1.3.0"
 click = "^7.0"
+prompt-toolkit = "<3.0.0"
 ptpython = "^2.0"
 jedi = "0.17.2"
 loguru = "^0.3.2"
@@ -21,26 +24,35 @@ fabric = "^2.4"
 msgpack = "^0.6.1"
 PyYAML = "^5.1"
 arrow = "^0.15.7"
-colorama = "^0.4.1"
 gevent = "1.4.0"
 greenlet = "0.4.16"
-faker = "^2.0"
-terminaltables = "^3.1"
 distro = "^1.4"
 libtmux = "^0.8.2"
 redis = "^3.3"
-dill = "^0.3.0"
-watchdog = "^0.9.0"
-GitPython = "^3.0"
-docker = "^4.2.0"
 jinja2 = "^2.11.1"
 psutil = "^5.7.0"
-pudb = "^2019.1"
-prompt-toolkit = "<3.0.0"
 bottle = "^0.12.18"
-Whoosh= "^2.7.4"
-pycparser = "^2.20"
-zipp = "^3.1.0"
+colorama = "^0.4.1"
+
+faker = { version = "^2.0", optional = true }
+terminaltables = { version = "^3.1", optional = true }
+dill = { version = "^0.3.0", optional = true }
+Whoosh = { version = "^2.7.4", optional = true }
+watchdog = { version = "^0.9.0", optional = true }
+GitPython = { version = "^3.0", optional = true }
+docker = { version = "^4.2.0", optional = true }
+
+[tool.poetry.extras]
+# core extras
+dill = ["dill"]
+whoosh = ["Whoosh"]
+# optional modules, mostly define __optional__ = True before export_module_as
+fake = ["faker"]
+terminaltable = ["terminaltables"]
+syncer = ["watchdog"]
+git = ["GitPython"]
+docker = ["docker"]
+all = ["dill", "Whoosh", "faker", "terminaltables", "watchdog", "GitPython", "docker"]
 
 [tool.poetry.dev-dependencies]
 ipython = "^7.6"
@@ -53,6 +65,8 @@ codecov = "^2.0"
 hypothesis = "^4.28"
 pytest-cov = "^2.7"
 objgraph = "^3.4.1"
+pdoc3 = "^0.6.3"
+pudb = "^2019.1"
 
 [tool.poetry.scripts]
 jsng = "jumpscale.entry_points.jsng:run"
